+++ hadoop/config/JanusGraphHadoopConfiguration.java
@@ -16,7 +16,7 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package grakn.core.graph.hadoop.config;
+package grakn.core.hadoop.config;
 
 import grakn.core.graph.diskstorage.Backend;
 import grakn.core.graph.diskstorage.configuration.ConfigNamespace;
