+++ hadoop/formats/cql/CqlBinaryRecordReader.java
@@ -16,7 +16,7 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package grakn.core.graph.hadoop.formats.cql;
+package grakn.core.hadoop.formats.cql;
 
 import com.datastax.oss.driver.api.core.cql.Row;
 import grakn.core.graph.diskstorage.Entry;
