+++ hadoop/formats/util/input/SystemTypeInspector.java
@@ -16,7 +16,7 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package grakn.core.graph.hadoop.formats.util.input;
+package grakn.core.hadoop.formats.util.input;
 
 
 public interface SystemTypeInspector {
